+++ Render/MapRenderer.js
@@ -38,6 +38,7 @@ class MapRenderer {
         // BACKROUND COLOR
         Renderer.drawRect(Renderer.color(renderContext.settings.mapBackgroundColor[0] ?? 0, renderContext.settings.mapBackgroundColor[1] ?? 0, renderContext.settings.mapBackgroundColor[2] ?? 0, renderContext.settings.mapBackgroundColor[3] ?? 150), x, y, size, size)// Background
 
+        // Render the map
         this.tabs[this.selectedTabIndex].draw(renderContext, dungeonMap, mouseX, mouseY)
 
         // THIS IS THE LEFT TOP AND RIGHT BORDER IN THAT ORDER
@@ -66,8 +67,8 @@ class MapRenderer {
         scoreInfoLore.push(`&f`)
         scoreInfoLore.push(`&fMin Secrets (s+): &a${scoreInfo.secretsFound ? scoreInfo.minSecrets : "?"}`)
         scoreInfoLore.push(`&fDeath penalty: &c${scoreInfo.deathPenalty}`)
-
-        if (renderContext.scoreInfoUnderMap === "simplified") {
+        if (renderContext.settings.scoreInfoUnderMap === "simplified") {
+            //ChatLib.chat("Got to line 71")
             let scoreInfoHeight = 10 * size / 100
             Renderer.drawRect(Renderer.color(renderContext.settings.extraInfoBackroundColor[0] ?? 0, renderContext.settings.extraInfoBackroundColor[1] ?? 0, renderContext.settings.extraInfoBackroundColor[2] ?? 0, renderContext.settings.extraInfoBackroundColor[3]), x, y + size, size, scoreInfoHeight)
 
@@ -95,7 +96,7 @@ class MapRenderer {
 
                 renderLore(mouseX, mouseY, scoreInfoLore)
             }
-        } else if (renderContext.scoreInfoUnderMap === "legalmap") {
+        } else if (renderContext.settings.scoreInfoUnderMap === "legalmap") {
             let scoreInfoHeight = 20 * size / 200
             Renderer.drawRect(Renderer.color(renderContext.settings.extraInfoBackroundColor[0] ?? 0, renderContext.settings.extraInfoBackroundColor[1] ?? 0, renderContext.settings.extraInfoBackroundColor[2] ?? 0, renderContext.settings.extraInfoBackroundColor[3]), x, y + size, size, scoreInfoHeight)