+++ Render/RenderContext.js
@@ -474,7 +474,8 @@ class RenderContext {
         customRoomColorUnknown = [64, 64, 64, 255],
         customRoomColorWitherDoor = [0, 0, 0, 255],
         customRoomGapSize = 9,
-        customDoorSize = 15
+        customDoorSize = 15,
+        spinnyMap = false
     }) {
         return {
             showMap,
@@ -536,7 +537,8 @@ class RenderContext {
             customRoomColorUnknown,
             customRoomColorWitherDoor,
             customRoomGapSize,
-            customDoorSize
+            customDoorSize,
+            spinnyMap
         }
     }